+++ source/Public/Start-SqlElasticJob.ps1
@@ -101,11 +101,11 @@ function Start-SqlElasticJob
 
         $startParameters = Add-OptionalParameter -Parameter $startParameters -BoundParameter $PSBoundParameters -Name 'Wait'
 
-        Write-PSFMessage -Level Verbose -Message ('Starting Elastic Job ''{0}'' on agent ''{1}''.' -f $Name, $AgentName) -Tag 'job', 'execution'
+        Write-PSFMessage -Level Output -Message ('Starting Elastic Job ''{0}'' on agent ''{1}''.' -f $Name, $AgentName) -Tag 'job', 'execution'
 
         $execution = Start-AzSqlElasticJob @startParameters
 
-        Write-PSFMessage -Level Verbose -Message ('Started Elastic Job ''{0}'' on agent ''{1}''.' -f $Name, $AgentName) -Tag 'job', 'execution'
+        Write-PSFMessage -Level Output -Message ('Started Elastic Job ''{0}'' on agent ''{1}''.' -f $Name, $AgentName) -Tag 'job', 'execution'
 
         return $execution
     }